Add an option to exclude trailing brackets in titles when searching
This covers #16 for searches of remote sources. We (currently) do not
support doing this for local files because it has significant potential
to be frustrating if the text in brackets is a distinguishing part of
the track title (since chances are that local sources are earlier in the
search sequence and so if you ignored the part in brackets you'd always
get the wrong lyrics in that case).