Add support for retrieving multiple lyrics from each source
Fixes #33.

This is not useful by itself because the normal search still
only actually uses the first result. This is important because it avoids
the need for us to make many, many queries to each website only to throw
away all of those results (possibly barring one). This avoids additional
web traffic as well as making searching faster.