Disable search avoidance for all remote tracks
An internet radio is always the same metadb_handle_ptr as far as fb2k
is concerned, but refers to different tracks. Since we store search
avoidance data on the metadb handle, we can't do it for internet radios
because then we might avoid searching for tracks that we'd otherwise
find just because some of the tracks on the radio don't have lyrics.