Support saving lyrics from internet radio streams to file
Since there are no track files available locally for internet radio
streams, we need to redirect attempts to save lyrics into their metadata
so that they instead go to a file somewhere. We can't just let it go
to the localfiles source and leave it at that, because the user might
have configured it to save to the track directory (which again, is not
something that we have available to us), so we just always go to the
config directory.