Add improved current-timestamp tracking for internet radio streams
fb2k does not reset the "current position" reported by the playback
controller when a new track starts on an internet radio stream.
We make the assumption that we only receive new dynamic metadata when
a new track starts, in order to be able to use that as our new baseline
timestamp.

We still don't seem to have any way of computing the length of a track
on an internet radio stream and so scrolling doesn't work for unsynced
lyrics. LyricShowPanel3 does manage to scroll, but seemingly for a
duration that was guessed, since it doesn't seem to actually line up
with the end of the track.

With the exception of the track length, this completes #49