Fix a case where NetEase sometimes returns garbage results unnecessarily
This is equivalent to PR #291 from HurakaMa. All credit to them for the
change. I'm just committing here because I wanted some minor tweaks and
it was simpler to just add them myself and make an equivalent one-line
change. Namely that PR did not include a changelog entry and I wanted
to have the IP used be trivially mappable to one of China's IP ranges
(they have 2022.96.0.0/12 according to https://ipinfo.space/CountryIPBlocks/CN_China.html)