Fix the lyric metadata index initialisation messing with search avoidance
The init callback classes had the same name and they were both in the
global namespace, so instead of one singleton for each, we'd just get
two singletons for one of them (which one is undefined behaviour).

We could have simply moved the init callbacks into an anonymous namespace
but instead I took the opportunity to pull that code out into a central
place next to the metadb index client and make  it impossible for this
problem to happen again. Also its now a bit less code to define a new
metadb index.