Update the lyric metadata index when track metadata changes
This ensures that if you get lyrics for a track, and then change any
of the metadata, we also update the metadata index so that it will still
show the historical metadata for that track.

We should probably also consider migrating lyrics stored in localfiles
so that they will continue to be found after the change.