Simplify musixmatch error handling
Leveraging the fact that if you ask cJSON for an object item and pass it
a nullptr object, it will short-circuit and just pass you another null
straight back. This combined with the fact that the IsString/Array/etc
functions will return false when the given item is null means we can
do a chain of queries and then just check that the resulting item is
the type that we need.