Fix metal-archives always returning a 403.
It would appear that metal-archives now rejects with a 403 any request
made via HTTP/1.1, but sensibly serves responses to requests made via
HTTP/2.

The foobar2000 SDK doesn't give us a way to make HTTP/2 requests, so
now we use libcurl for this instead, which gives us much more flexibility
anyway.

This serves as an experimental replacement for all network calls, which
we could migrate  over to using libcurl in future if there was any need
or benefit to it. It would now at least be easy.

Fixes #457, #481